[M3] Fix some class names and locators
