+++ Epam.ReportPortal.Automation/Epam.ReportPortal.Automation.UiiBusinessLayer/WebObjects/Components/FooterComponent.cs
@@ -3,17 +3,19 @@ using OpenQA.Selenium;
 
 namespace Epam.ReportPortal.Automation.UiiBusinessLayer.WebObjects.Components;
 
-public class LeftPanelComponent : BaseWebComponent
+public class FooterComponent : BaseWebComponent
 {
     public override IWebElement Root => Driver.FindElement(By.CssSelector("TODO"));
 
-    public IWebElement DashboardsItem => Driver.FindElement(By.XPath("//div[contains(@class,'sidebarButton__sidebar-nav-btn') and .//span[text()='Dashboards']]"));
+    public IWebElement ApplicationVersionLabel => Driver.FindElement(By.CssSelector("TODO"));
 
-    public IWebElement LaunchesItem => Driver.FindElement(By.XPath("//div[contains(@class,'sidebarButton__sidebar-nav-btn') and .//span[text()='Launches']]"));
+    public IWebElement GitHubLink => Driver.FindElement(By.CssSelector("TODO"));
 
-    public IWebElement FiltersItem => Driver.FindElement(By.CssSelector("TODO"));
+    public IWebElement SlackLink => Driver.FindElement(By.CssSelector("TODO"));
 
-    public IWebElement ProjectSettingsItem => Driver.FindElement(By.CssSelector("TODO"));
+    public IWebElement ContactUsLink => Driver.FindElement(By.CssSelector("TODO"));
 
-    public IWebElement UserIcon => Driver.FindElement(By.CssSelector("TODO"));
+    public IWebElement DocumentationLink => Driver.FindElement(By.CssSelector("TODO"));
+
+    public IWebElement PolicyLink => Driver.FindElement(By.CssSelector("TODO"));
 }
